Enable tracking of the warfile version
One way to deploy new apps is to issue a new warfile. These changes to
the module allow Puppet to automatically deploy new app versions based
on nothing but the warfile. This is achieved by seting replace=true on
the staging::file, exposed through an option to tomcat::war called
"replace".